Wrote a new I2C driver.
I am trying to figure out this I2C issue. It keeps on randomly tanking on me and reporting bad values, I am not sure why. So I decided to eliminate the MicroPython mechanics involved in it. There could be something in there that is causing the problem.

I was going to change to ESP-IDF 5.1.2 but there is no large differences that jump right out at me compared to 5.0.6. 5.2 is where they released the whole new API for I2C and IO know that it is not going to be something easy to migrate to. I may try it in the upcoming weeks to see what is needed.