adds --octal-flash and --flash-size for an ESP32_GENERIC_S3 build
There are variants of the esp32-s3 that have 4mb, 8mb, 16mb and 32mb flash and that is what using --flash-size=? allows us to set.

The --octal-flash you have to add if the flash uses octal SPI. This is seen with the 32mb flash size. There are some changes that need to occur in order for the flash to work properly. So if you have an esp32-s3 N32R8 you will want to use the following build command

```
make.py esp32 clean submodules mpy_cross BOARD=ESP32_GENERIC_S3 BOARD_VARIANT=SPIRAM_OCT --octal-flash --flash-size=32
```