updates SPIBus to accept `spi.Bus` instance
Changes the XPT2046 driver to use `spi.Bus` and `spi.Device`
Adds use of macro to enable logging. I still have to update the build script to turn the debug logging on and off from the command line.

I also need to write new spi handling for the rest of the MCU's