adds a more refined touch screen calibration.
The old version of the touch screen calibration didn't handle the touch screen being slightly rotated from the display. This new version takes that into consideration and will adjust the touch points to compensate for it. There is also an updated calibration GUI as well. The GUI needs some more work to add additional information on what to do. It's pretty easy to understand. I do need to add a timer that will time out and reset the calibration if the user is not able to click the OK button. I also need to add a cancel button to exit the calibration as well.