reverts some of the changes made to the ili9341 driver.
I had hard coded `color_space` to BGR when it turns out that there are variants that use RGB. So I am defaulting `color_space` to RGB which is what is done across the board for all other display drivers. The users that have the BGR type displays IC's will need to set this parameter.

Fixes #55