Gives a speed boost to collecting the submodules
Since all of the submodules are only there to provide files and not for the purposes of being able to update them using git there is no reason to clone the full depth of the submodules repositories. The updates that are done now are shallow updates and only include 1 level of commit data. This has greatly improved the in which it takes to collect the submodules.