fixes display rotation for RGB displays.
RGB displays do not have hardware rotation so software rotation must be used. This specific feature seems to have been removed from LVGL 9.x so I have come up with a way to work around the issue by rotating the layers which does the same thing without needing to write a custom flush function to handle reorderinmg the pixels. I am hoping this is going to be a faster way or handling the software rotation.