adds dual, quad and octal data transmitting modes for the SPI LCD and also for SPI Device.
you are able to "mix and match" when using the different type on a bus.

Say you have a quad display and a normal spi touch panel. They can both be placed on the same bus. When constructing the display bus you add `quad=True` when constructing the `lcd_bus.SPIBus` instance. The display would get connected to miso, mosi, data2 and data3 pins and the touch would only get connected to the miso and mosi pins.