adds `enable_input_priority` method to touch drivers.
What this feature does is it places priority on user input by looping the read process and updating the display after the input has been processed. The way the indev works in LVGL is it will read touch input every 33 milliseconds but it doesn't update the display right away. It lets the refresh timer take care of that which once again occurs every 33 milliseconds. This means it is possible for user input to become realized on the display 33 milliseconds after it has occurred. This type of behavior doesn't lead to a very smooth UI. It becomes even worse if the user code in the main loop takes a long time to complete. That compounds the issue. By using this feature it will place user input at the highest priority, it will not allow the main loop to continue until there is no more user input.

Events that occur as the result of user input are ran as they would be when the input occurs.