adds support for MIPI DSI Bus
This is in preparation for when MicroPython is going to support using version 5.3.x of the ESP-IDF.
This is only going to be available using the ESP32-P4 series of SOC's. This is one feature I can't wait to have available.

I believe that supporting ESP-IDF 5.3.x requires quite a bit of modification to MicroPython which is the reason why I have added support for it. I believe we are getting pretty close to another release of MicroPython and that release hopefully will have support for ESP-IDF 5.3.x.

One thing I noticed is that the driver in the ESP-IDF is very similar to the RGB driver, I am hoping that the P4 has faster memory than the Sx line of SOC's. The memory speed is a HUGE bottleneck when dealing with displays that do not have any GRAM. While it is fantastic that the P4 supports MIPS DSI if the memory speed has not been improved then the speed of MIPI DSI is never going to be realized. It will still be better to use than RGB due to the low pin count but I8080 will still be preferred because of the display having GRAM and being able to have partial frame buffers.

I have to dive into the datasheet for the P4 to find out.