fixes repl for unix and macOS build
There was a change I had made that would not allow more advanced things like history to be available when using the repl. I have fixed this issue.
I also started working on adding a roll back to restore all micropython code back to it's originals upon build failure and also when a build completes sucessfully. This allows for keeping micropython scrubbed of any changes in the event a user wants to compile for a different port. While I am not modifying anything outside of the the port directories at the moment the need might arrise at some point in time.