adds using a toml file for the build.
I am going to have to spend some time to write the documentation for this feature. It can be a very powerful tool for people to use when compiling. It allows for supplying all of the build commands and it will also build a startup file for a display and touch.

People asked for it and now they have it. So anyone that has a working display and wants to provide me with their startup code I would be more than happy to start adding toml files that are specific to already made display boards.

The example see in the root folder is for the WT32-SC01-Plus. Well most of it is. I don't believe I have the MCU portion of the toml correct and I will need to update that.