Fixes RGB rotation..
Thank you to @getznerh for taking the time to track down the issue with this. I have made some many changes to this driver trying to improve the performance of it and I didn't get the calculations quite right. TY for fixing this for us. This is a big help and now that it is fixed I am going to utilize the rotation across all of the bus drivers.