adds new c module that exposes FreeRTOS to Python.
This is for advanced users that know how to deal with semaphores and threads. It enables true parallel processing allowing use of both cores.
It is only available for the ESP32-S3 for the time being and I will make it available to all dual core ESP32 MCU's. To enable using it you need to add `--py-freertos` to your build command.

**WARNING**
Using this disables the Python GIL so care much be given to control access to objects by multiple threads using semaphores and queues otherwise memory corruption will occur.