working on the new binding code generator script.
I am going to flatten the API so it matches the C API. It drives me bonkers having to hunt for things. I am hoping the end result will be less memory use, smaller compiled binary size and faster. I know that is going to be a bit of a stretch to nail all 3 of them but we will see what kind of magic I am able to do. I am going to have to really look at those resources and see what the tradeoffs will be. Typically with speed you have more memory use so we will see what happens there. I believe just flattening the API is going to speed things up and also reduce memory because the lookup tables for methods and things of that nature will not need to be there.