Reorganization of the lcd_bus extension module source and header files.
This needs to be done in order to handle the rotation due to the rotation running on a second core if applicable. We have to separate the MCU's so we have the ability to use the appropriate RTOS that is available for the port being used.

It is also a cleaner way to understand what is happening in general with the build. We have the ability to see what is and is not available. This change also adds all busses to all ports and the busses that are not supported will throw an error stating they are not available. This would be a good mechanism to use when determining how to set up the display driver based on the MCU being used. Simple exception catching when starting the display bus will allow the program to determine what board to use. A fantastic use of this is when writing the software on a desktop using the unix port and the final program is going to be run on the ESP32 port.