Working on reorganizing the lcd bus drivers
The remodel of the bus drivers is to allow for better handling of each of the bus types on a per MCU/board basis.
I am also expanding the software rotation in the RGB bus driver to allow its use across all bus types.
I have added RGB888 dithering and RGB565 dithering that will be able to be used with or without rotation. rgb565 byte swap is being added in a manner that will allow only a single iteration over the frame buffer data is either software rotation is being used or dithering is being used or both.

For memory constrained ESP32 boards I have to work out a way to handle the transmitting of the frame buffer without needing to allocate additional buffers. Ideally it is going to be best to create smaller buffers to allow the allocation of at least 1 additional frame buffer. Ideally if 2additional frame buffers are able to be allocated, in DMA memory space it would be best so while one buffer is transmitting the data is able to be copied into the second buffer. If only 1 buffer is able to be allocated, then None of the buffers will need to be allocated in DMA memory. This will at the very least give a performance boost because LVGL is going to be able to render to one of the user supplied buffers without being interrupted at all from ISRs during the sending to a display. This is because the sending to the display is going to be done on the second core even if 1 buffer is used for sending and that buffer is not in DMA memory. I need to put some thought into how to handle the buffer allocation in the best way as I don't want to add anything that can be confusing for the user. If everything works like I am hoping things should actually get a little easier to use because the user is not going to be given the ability to specify the type of memory a buffer has to be allocated in. The user will be able to simply create bytearrays for the frame buffers and pass memoryview instances pointing to those bytearrays to the display driver.

The CI is going to fail for this commit. The code is not finished yet.